# modules/development-host/src/main/resources/site/addzero/studio/development/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Console Client API",
    "version": "1.0.0",
    "description": "Generated from shared transport contracts."
  },
  "paths": {
...
    "/studio/api/lowcode/library/validate": {
      "post": {
        "operationId": "validateLibrary",
        "tags": [
          "Library"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/MetadataValidationResult"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/studio/api/lowcode/library/delete": {
      "delete": {
        "operationId": "deleteLibraries",
        "tags": [
          "Library"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": [
                        "boolean",
                        "null"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/studio/api/lowcode/library-feature/validate": {
      "post": {
        "operationId": "validateLibraryFeature",
        "tags": [
          "Library Feature"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryFeatureCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/MetadataValidationResult"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/studio/api/lowcode/library-feature/delete": {
      "delete": {
        "operationId": "deleteLibraryFeature",
        "tags": [
          "Library Feature"
        ],
        "summary": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": [
                        "boolean",
                        "null"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/studio/api/lowcode/model/validate": {
      "post": {
        "operationId": "validateModel",
        "tags": [
          "Model"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/MetadataValidationResult"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/studio/api/lowcode/model": {
      "delete": {
        "operationId": "deleteModels",
        "tags": [
          "Model"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": [
                        "boolean",
                        "null"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/studio/api/lowcode/dto/validate": {
      "post": {
        "operationId": "validateDto",
        "tags": [
          "DTO"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/MetadataValidationResult"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/studio/api/lowcode/dto": {
      "delete": {
        "operationId": "deleteDtos",
        "tags": [
          "DTO"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": [
                        "boolean",
                        "null"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/studio/api/lowcode/dto/reuse-analysis": {
      "post": {
        "operationId": "analyzeDtoReuse",
        "tags": [
          "DTO"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/JsonElement"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/studio/api/lowcode/convention-file/validate": {
      "post": {
        "operationId": "validateConventionFile",
        "tags": [
          "Convention File"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConventionFileCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/MetadataValidationResult"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/studio/api/lowcode/convention-file": {
      "delete": {
        "operationId": "deleteConventionFiles",
        "tags": [
          "Convention File"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": [
                        "boolean",
                        "null"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/studio/api/lowcode/constant/validate": {
      "post": {
        "operationId": "validateConstant",
        "tags": [
          "Constant"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstantCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/MetadataValidationResult"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/studio/api/lowcode/constant/save": {
      "post": {
        "operationId": "saveConstant",
        "tags": [
          "Constant"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstantCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ConstantView"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/studio/api/lowcode/constant": {
      "delete": {
        "operationId": "deleteConstants",
        "tags": [
          "Constant"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": [
                        "boolean",
                        "null"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "JsonElement": {
        "title": "JsonElement",
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "type": "null"
          },
          {
            "type": "array"
          },
          {
            "type": "object",
            "additionalProperties": true
          }
        ]
      },
...
      "LibraryCommand": {
        "type": "object",
        "title": "LibraryCommand",
        "required": [
          "code",
          "displayName",
          "spec"
        ],
        "properties": {
          "id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "code": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          },
          "status": {
            "type": "integer"
          },
          "spec": {
            "$ref": "#/components/schemas/LibrarySpec"
          }
        }
      },
      "MetadataValidationResult": {
        "type": [
          "object",
          "null"
        ],
        "title": "MetadataValidationResult",
        "required": [
          "valid"
        ],
        "properties": {
          "valid": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "LibraryFeatureCommand": {
        "type": "object",
        "title": "LibraryFeatureCommand",
        "required": [
          "libraryId",
          "featureCode",
          "name"
        ],
        "properties": {
          "id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "libraryId": {
            "type": "integer"
          },
          "parentId": {
            "type": [
              "integer",
              "null"
            ]
          },
          "featureCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ModelCommand": {
        "type": "object",
        "title": "ModelCommand",
        "required": [
          "featureId",
          "modelCode",
          "name",
          "className",
          "tableName"
        ],
        "properties": {
          "id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "featureId": {
            "type": "integer"
          },
          "modelCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "packageName": {
            "type": "string"
          },
          "className": {
            "type": "string"
          },
          "tableName": {
            "type": "string"
          },
          "modelType": {
            "type": "string",
            "enum": [
              "ENTITY",
              "MAPPED_SUPERCLASS",
              "EMBEDDABLE"
            ]
          },
          "status": {
            "type": "integer"
          },
          "version": {
            "type": "integer"
          },
          "contributorId": {
            "type": [
              "string",
              "null"
            ]
          },
          "entityConfig": {
            "$ref": "#/components/schemas/EntityConfigCommand"
          },
          "routeConfig": {
            "$ref": "#/components/schemas/RouteCommand"
          },
          "remark": {
            "type": [
              "string",
              "null"
            ]
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldCommand"
            }
          },
          "queries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryCommand"
            }
          },
          "relations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelationCommand"
            }
          }
        }
      },
...
    }
  }
}